bot/requirements.txt
aiogram
python-dotenv
captcha
fastapi[standard]
